third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-history-go-0.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
      assert_equals(e.canRespond, true);
      assert_equals(e.hashChange, false);
     assert_equals(e.formData, null);
      e.preventDefault();
    });

    i.contentWindow.history.go(0);
  });
}, "history.go(0) fires the navigate event");
</script>
